For Exercise 1 , I have added the following files: .dockerignore Cargo.lock Cargo.toml Dockerfile -SRC ---main.rs   -> contains a hello word rust application , the dockerfile will build this application
For Exercise 2 , I have added the following files:

deploy.sh -> contains a script which will initializ terraform , plan and apply the infrastrcture , it will also add the EKS cluster context ,then deploy the application from the deploy.yaml file

deploy.yaml -> this file contains a kuberentes namespace, a kubernetes deployment and a service type NodePort . The deployment uses the image which we have built in Exercise 1 , the image is deployed to a public docker hub repository

Dockerfile -> This file is used to build a docker image which contains all the tools and dependecines needed to run the deploy.sh file

els-user-date.sh -> contains eks user data - bootstrapping the EKS cluster

main.tf -> contains the IaC to be build via terraform.

manage_lb_target_group_attachment.sh -> contains a function which will comment/uncomment the last resource in main.tf , it has a foreach statement that will fail if the other resources are not applied first , so we first apply everything , while we have a block of code commented out by this script , then we uncomment the block of code using the script and apply everything again , the last apply will only add the missing resource.